Split scribe.py: extract classify.py (source bucketing)
Phase-0 refactor (#306), commit 3 of 5. Behaviour-preserving.

New bartleby/ingest/classify.py owns turning the supplied paths into
ingestible (file, ext) pairs and bucketing them against DB state:
_collect_files, _hash_file, _is_complete, _ResumeItem, and _classify
(skip / resume / queue / in-run-duplicate). All main-process work; only the
queued ParseRequest bucket crosses to the parse workers.

scribe.py imports `from bartleby.ingest import classify` and calls
classify._collect_files / classify._classify; the now-orphaned hashlib,
IMAGE_EXTENSIONS, and resolve_extension imports are dropped. Tests retargeted
to classify._collect_files / classify._is_complete.
