+++ tests/test_scribe.py
@@ -15,6 +15,7 @@ import bartleby.db.connection
 import bartleby.project
 from bartleby.commands import scribe
 from bartleby.ingest import parsers
+from bartleby.ingest import classify
 from bartleby.db.connection import open_db
 from bartleby.db.schema import EMBEDDING_DIM
 from bartleby.ingest.chunk import resolve_extension
@@ -1330,7 +1331,7 @@ def test_is_complete_text_document(isolated_project, tmp_path, mock_embed):
             "SELECT document_id FROM documents"
         ).fetchone()[0]
         writer = scribe.Writer(conn)
-        assert scribe._is_complete(writer, doc_id)
+        assert classify._is_complete(writer, doc_id)
     finally:
         conn.close()
 
@@ -1362,7 +1363,7 @@ def test_is_complete_false_when_image_uncaptioned(isolated_project):
             (doc_id, image_id, None, 0),
         )
         writer = scribe.Writer(conn)
-        assert not scribe._is_complete(writer, doc_id)
+        assert not classify._is_complete(writer, doc_id)
         assert [pi.image_id for pi in writer.uncaptioned_images(doc_id)] == [image_id]
 
         # Caption it → complete, and no longer pending.
@@ -1370,7 +1371,7 @@ def test_is_complete_false_when_image_uncaptioned(isolated_project):
             "UPDATE images SET analysis_json = '{}', analysis_model = 'm' "
             "WHERE image_id = ?", (image_id,),
         )
-        assert scribe._is_complete(writer, doc_id)
+        assert classify._is_complete(writer, doc_id)
         assert writer.uncaptioned_images(doc_id) == []
     finally:
         conn.close()
@@ -1891,7 +1892,7 @@ def test_resolve_format_filter_rejects_unknown_name():
 
 def test_collect_files_single_extensionless_pdf(tmp_path):
     pdf = _text_pdf(tmp_path / "docket_no_ext")
-    sources, unidentified = scribe._collect_files([pdf])
+    sources, unidentified = classify._collect_files([pdf])
     assert sources == [(pdf, ".pdf")]
     assert unidentified == []
 
@@ -1900,7 +1901,7 @@ def test_collect_files_single_unsupported_raises(tmp_path):
     p = tmp_path / "mystery"
     p.write_bytes(b"\x00\x01\x02\x03")
     with pytest.raises(ValueError):
-        scribe._collect_files([p])
+        classify._collect_files([p])
 
 
 def test_collect_files_directory_sniffs_and_reports_unidentified(tmp_path):
@@ -1911,7 +1912,7 @@ def test_collect_files_directory_sniffs_and_reports_unidentified(tmp_path):
     junk = d / "notes"                                      # unidentifiable
     junk.write_bytes(b"\x00\x01\x02\x03")
 
-    sources, unidentified = scribe._collect_files([d])
+    sources, unidentified = classify._collect_files([d])
 
     assert (pdf, ".pdf") in sources
     assert (named, ".pdf") in sources
@@ -1926,7 +1927,7 @@ def test_collect_files_unions_multiple_directories(tmp_path):
     pdf_a = _text_pdf(a / "a.pdf")
     pdf_b = _text_pdf(b / "b.pdf")
 
-    sources, unidentified = scribe._collect_files([a, b])
+    sources, unidentified = classify._collect_files([a, b])
 
     files = {p for p, _ in sources}
     assert files == {pdf_a, pdf_b}
@@ -1939,7 +1940,7 @@ def test_collect_files_dedupes_overlapping_roots(tmp_path):
     pdf = _text_pdf(d / "report.pdf")
 
     # The same file reachable both as an explicit file and under its directory.
-    sources, _ = scribe._collect_files([pdf, d])
+    sources, _ = classify._collect_files([pdf, d])
     assert sources.count((pdf, ".pdf")) == 1
     assert len(sources) == 1
 
@@ -1952,7 +1953,7 @@ def test_collect_files_only_filter_restricts_by_resolved_type(tmp_path):
     png = d / "image.png"
     png.write_bytes(_png_bytes())
 
-    sources, unidentified = scribe._collect_files([d], only={".pdf"})
+    sources, unidentified = classify._collect_files([d], only={".pdf"})
 
     assert sources == [(pdf, ".pdf")]
     # txt and png are intentional exclusions — not lumped into unidentified.
@@ -1967,7 +1968,7 @@ def test_collect_files_only_filter_matches_sniffed_type(tmp_path):
     sniffed = _text_pdf(d / "ATTACHMENT 7 - LOADING FORMS")  # extensionless PDF
     _write_txt(d / "notes.txt", "plain text")
 
-    sources, _ = scribe._collect_files([d], only={".pdf"})
+    sources, _ = classify._collect_files([d], only={".pdf"})
 
     # The content-sniffed PDF is kept by `--only pdf` just like a named one.
     assert sources == [(sniffed, ".pdf")]
